+++ tests/testthat/test02_blob.R
@@ -225,6 +225,103 @@ test_that("AAD authentication works",
 })
 
 
+test_that("Invalid transfers handled correctly",
+{
+    ## nonexistent endpoint
+    badname <- paste0(sample(letters, 20, TRUE), collapse="")
+    endp <- blob_endpoint(sprintf("https://%s.blob.core.windows.net", badname), key="foo")
+    cont <- blob_container(endp, "nocontainer")
+
+    # uploading
+    expect_error(upload_blob(cont, "../resources/iris.csv", "iris.csv"))
+
+    json <- jsonlite::toJSON(iris)
+    con <- textConnection(json)
+    expect_error(upload_blob(cont, con, "iris.json"))
+
+    rds <- serialize(iris, NULL)
+    con <- rawConnection(rds)
+    expect_error(upload_blob(cont, con, "iris.rds"))
+
+    # downloading
+    expect_error(download_blob(cont, "nofile", tempfile()))
+    expect_error(download_blob(cont, "nofile", NULL))
+
+    con <- rawConnection(raw(0), "r+")
+    expect_error(download_blob(cont, "nofile", con))
+
+
+    ## nonexistent container
+    bl <- stor$get_blob_endpoint()
+    cont <- blob_container(bl, "nocontainer")
+
+    # uploading
+    expect_error(upload_blob(cont, "../resources/iris.csv", "iris.csv"))
+
+    json <- jsonlite::toJSON(iris)
+    con <- textConnection(json)
+    expect_error(upload_blob(cont, con, "iris.json"))
+
+    rds <- serialize(iris, NULL)
+    con <- rawConnection(rds)
+    expect_error(upload_blob(cont, con, "iris.rds"))
+
+    # downloading
+    expect_error(download_blob(cont, "nofile", tempfile()))
+    expect_error(download_blob(cont, "nofile", NULL))
+
+    con <- rawConnection(raw(0), "r+")
+    expect_error(download_blob(cont, "nofile", con))
+
+
+    ## bad auth
+    bl$key <- "badkey"
+
+    cont <- blob_container(bl, "nocontainer")
+
+    # uploading
+    expect_error(upload_blob(cont, "../resources/iris.csv", "iris.csv"))
+
+    json <- jsonlite::toJSON(iris)
+    con <- textConnection(json)
+    expect_error(upload_blob(cont, con, "iris.json"))
+
+    rds <- serialize(iris, NULL)
+    con <- rawConnection(rds)
+    expect_error(upload_blob(cont, con, "iris.rds"))
+
+    # downloading
+    expect_error(download_blob(cont, "nofile", tempfile()))
+    expect_error(download_blob(cont, "nofile", NULL))
+
+    con <- rawConnection(raw(0), "r+")
+    expect_error(download_blob(cont, "nofile", con))
+
+    close(con)
+})
+
+
+test_that("chunked downloading works",
+{
+    bl <- stor$get_blob_endpoint()
+    cont <- create_blob_container(bl, "chunkdl")
+
+    orig_file <- "../resources/iris.csv"
+    new_file <- tempfile()
+    upload_blob(cont, orig_file, "iris.csv")
+
+    download_blob(cont, "iris.csv", new_file, overwrite=TRUE, blocksize=100)
+    expect_identical(readBin(orig_file, "raw", n=1e5), readBin(new_file, "raw", n=1e5))
+
+    con <- rawConnection(raw(0), open="r+")
+    download_blob(cont, "iris.csv", con, blocksize=130)
+    expect_identical(readBin(orig_file, "raw", n=1e5), readBin(con, "raw", n=1e5))
+
+    con <- download_blob(cont, "iris.csv", NULL, blocksize=150)
+    expect_identical(readBin(orig_file, "raw", n=1e5), readBin(con, "raw", n=1e5))
+})
+
+
 teardown(
 {
     bl <- stor$get_blob_endpoint()